refactoring (folders, functions, variables in App.js) and some styles tweaks.
